Thermo-nuclear: extract OpenCode billing into billing.rs
